Rename `i386` arch directory to `i686` Update all necessary comments and docs too
