+++ kernel/arch/Arch.hpp
@@ -13,8 +13,8 @@
 #include <meta/compiler.hpp>
 #include <stddef.h>
 #include <stdint.h>
-#if defined(__i386__)
-#    include <arch/i386/Arch.i386.hpp>
+#if defined(__i686__)
+#    include <arch/i686/Arch.i686.hpp>
 #endif
 
 /**
