+++ kernel/arch/i686/asm/flush.S
