+++ kernel/arch/i686/asm/tasks.S
