+++ kernel/arch/i686/boot/boothelp.cpp
