+++ kernel/arch/i686/boot/panic.cpp
