+++ kernel/arch/i686/boot/stivale.cpp
