+++ kernel/arch/i686/boot/stivale2.S
