+++ kernel/arch/i686/linker.ld
