+++ kernel/arch/i686/regs.hpp
