+++ kernel/sys/tasks.cpp
@@ -17,7 +17,7 @@
 #include <dev/serial/rs232.hpp>
 #include <stdint.h>
 #include <x86gprintrin.h>   // needed for __rdtsc
-#include <arch/i386/timer.hpp> // TODO: Remove ASAP
+#include <arch/i686/timer.hpp> // TODO: Remove ASAP
 
 /* forward declarations */
 static void _enqueue_task(struct tasklist *, task *);
